+++ build.gradle
@@ -18,6 +18,6 @@ dependencies {
 
 test {
     useTestNG() {
-        suites "src/test/resources/HW20.xml"
+        suites "src/test/resources/HW21.xml"
     }
 }
